+++ src/functionalTest/resources/application.yaml
@@ -78,7 +78,7 @@ pba:
 feature-toggle:
   toggle:
     example: ${FEATURE_EXAMPLE:false}
-    bundleByCategoryEnabled: false
+    bundleByCategoryEnabled: ${BUNDLE_BY_CATEGORY_ENABLED:false}
 
 xui:
   url: ${XUI_URL:https://manage-case.aat.platform.hmcts.net/cases/case-details}
