+++ src/integrationTest/resources/application.yaml
@@ -76,6 +76,7 @@ pba:
 feature-toggle:
   toggle:
     example: ${FEATURE_EXAMPLE:false}
+    bundleByCategoryEnabled: ${BUNDLE_BY_CATEGORY_ENABLED:false}
 
 xui:
   url: ${XUI_URL:https://manage-case.aat.platform.hmcts.net/cases/case-details}
